move TransposeScheduler from registry.cpp to transpose.cpp
