+++ csrc/scheduler/transpose.cpp
@@ -5,28 +5,129 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 // clang-format on
-#include <scheduler/transpose.h>
-
-#include <debug.h>
-#include <device_lower/utils.h>
-#include <executor_utils.h>
-#include <inlining.h>
-#include <instrumentation.h>
-#include <ir/iostream.h>
-#include <ir/utils.h>
-#include <options.h>
-#include <scheduler/pointwise_utils.h>
-#include <scheduler/registry.h>
-#include <scheduler/utils.h>
-#include <scheduler/vectorize_helper.h>
-#include <transform_replay.h>
-#include <utils.h>
 
 #include <ATen/cuda/CUDAContext.h>
-
-#include <algorithm>
+#include <debug.h>
+#include <inlining.h>
+#include <instrumentation.h>
+#include <scheduler/debug_utils.h>
+#include <scheduler/reduction_utils.h>
+#include <scheduler/registry_utils.h>
+#include <scheduler/transpose.h>
+#include <scheduler/utils.h>
+#include <scheduler/vectorize_helper.h>
 
 namespace nvfuser {
+
+TransposeScheduler::TransposeScheduler(
+    Fusion* fusion,
+    SchedulerRuntimeInfo& runtime_info,
+    HeuristicSummary* data_cache)
+    : SchedulerEntry(ScheduleHeuristic::Transpose) {
+  computeHeuristics(fusion, runtime_info, data_cache);
+}
+
+bool TransposeScheduler::canScheduleCompileTime(Fusion* fusion) {
+  // Check that inputs of all select/gather-like ops are fusion inputs
+  if (registry_utils::rejectScheduleForMemoryPromotion(
+          fusion, ScheduleHeuristic::Transpose)) {
+    return false;
+  }
+
+  // Fusions handled by transpose scheduler cannot have MmaOp.
+  if (!ir_utils::getMmaOps(fusion).empty()) {
+    scheduler_debug_utils::canScheduleRejectReason(
+        ScheduleHeuristic::Transpose, "no support for mma ops.");
+    return false;
+  }
+
+  for (auto select : ir_utils::getSelectOps(fusion)) {
+    auto root = TensorDomain::noReductions(
+        select->input(0)->as<TensorView>()->getMaybeRFactorDomain());
+    if (select->getIndexedID() == root[root.size() - 1]) {
+      scheduler_debug_utils::canScheduleRejectReason(
+          ScheduleHeuristic::Transpose,
+          "SelectOp on inner dim is not supported by transpose scheduler yet."
+          "In transpose scheduler, we want to leave the select dim alone, instead of creating a tile for it.");
+      return false;
+    }
+  }
+  for (auto idx_sel : ir_utils::getIndexSelectOps(fusion)) {
+    auto root = TensorDomain::noReductions(
+        idx_sel->input(0)->as<TensorView>()->getMaybeRFactorDomain());
+    if (idx_sel->getIndexedID() == root[root.size() - 1]) {
+      scheduler_debug_utils::canScheduleRejectReason(
+          ScheduleHeuristic::Transpose,
+          "IndexSelectOp on inner dim is not supported by transpose scheduler yet."
+          "In transpose scheduler, we want to leave the select dim alone, instead of creating a tile for it.");
+      return false;
+    }
+  }
+  for (auto torch_gather : ir_utils::getTorchGatherOps(fusion)) {
+    auto root = TensorDomain::noReductions(
+        torch_gather->input(0)->as<TensorView>()->getMaybeRFactorDomain());
+    if (torch_gather->dim() == (int)root.size() - 1) {
+      scheduler_debug_utils::canScheduleRejectReason(
+          ScheduleHeuristic::Transpose,
+          "TorchGatherOp on inner dim is not supported by transpose scheduler yet."
+          "In transpose scheduler, we want to leave the select dim alone, instead of creating a tile for it.");
+      return false;
+    }
+  }
+
+  if (!hasAtLeastTwoValidGroups(fusion)) {
+    scheduler_debug_utils::canScheduleRejectReason(
+        ScheduleHeuristic::Transpose,
+        "cannot find two mismatching inner most dimensions");
+    return false;
+  }
+
+  auto reduction_ops = ir_utils::getReductionOps(fusion);
+
+  if (!reduction_ops.empty()) {
+    scheduler_debug_utils::canScheduleRejectReason(
+        ScheduleHeuristic::Transpose, "no support for reduction ops");
+    return false;
+  }
+
+  if (registry_utils::hasNonUniqueBcast(fusion)) {
+    scheduler_debug_utils::canScheduleRejectReason(
+        ScheduleHeuristic::Transpose,
+        "Broadcasting dimension might be broadcasting to multiple sizes.");
+    return false;
+  }
+
+  return true;
+}
+
+bool TransposeScheduler::canScheduleRunTime(
+    Fusion* fusion,
+    SchedulerRuntimeInfo& runtime_info,
+    HeuristicSummary* data_cache) {
+  FUSER_PERF_SCOPE("TransposeScheduler::canScheduleRunTime");
+
+  auto reason =
+      getTransposeRuntimeRejectReason(fusion, data_cache, runtime_info);
+  if (!reason.empty()) {
+    scheduler_debug_utils::canScheduleRejectReason(
+        ScheduleHeuristic::Transpose, reason);
+    return false;
+  }
+  return true;
+}
+
+void TransposeScheduler::schedule(Fusion* fusion) {
+  FUSER_PERF_SCOPE("Schedule Transpose Fusion");
+  scheduleTranspose(fusion, transposeParams());
+}
+
+void TransposeScheduler::computeHeuristics(
+    Fusion* fusion,
+    SchedulerRuntimeInfo& runtime_info,
+    HeuristicSummary* data_cache) {
+  params_ = getTransposeHeuristics(fusion, runtime_info, data_cache);
+  NVF_ERROR(params_ != nullptr);
+}
 
 namespace {
